Move oot fragment to fragment package
